+++ src/klib/serial.h
@@ -6,13 +6,14 @@
  */
 
 #include <stdint.h>
+#include <stddef.h>
 
 #ifndef SERIAL_H_
 #define SERIAL_H_
 
 #define COM1 0x3f8
 
-void init_serial(uint16_t port);
+void init_serial(size_t port);
 int is_transmit_empty(uint16_t port);
 void write_serial(uint16_t port, char a);
 void print_serial(uint16_t port, const char* s);
